added retry mechanism for registry apis and fixed a couple of issues
