+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "evergreen-org-crawler",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "",
   "main": "./build/index.js",
   "types": "./build/index.d.ts",
